JM pointed out.. Here's the exact quote:
*Important*
When you send an email to multiple recipients (recipients are "To", "CC", and "BCC" addresses) and the call to Amazon SES fails, the entire email is rejected and none of the recipients will receive the intended email. For example, if a "BCC" address is on the blacklist, then the call will fail and the email is not sent to any of the "To", "CC", or "BCC" addresses. *We therefore recommend that you send an email to one recipient at a time.*

http://docs.aws.amazon.com/ses/latest/DeveloperGuide/sending-email.html

git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtPlatform/trunk@18869 3dfb626b-079f-4ab1-a9f6-4654e5377508