First part of refactoring the landing page.
The app is now under a prefix (default /app) and /landing/whatever is
a redirect to /whatever; if there is a landing page, it is served as
root, otherwise root is a redirect to /login.

Because the landing page has no "go to the app" button, and/or the loginapp
doesn't send you there if you are already logged in (it makes you log in again),
if you appear to be logged in the root is a redirect to /app. The problem with that
is that you cannot get back to the root content without explicitly putting
/index.html on the end...



git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtPlatform/trunk/nti.dataserver@36808 3dfb626b-079f-4ab1-a9f6-4654e5377508