Support for (and use) gunicorn 19.3.
The main observable difference is that the PROXY protocol is now /required/ in production environments. There may be some minor changes to log formats.
The QUIT and TERM signals are now reversed, which should make it play better with supervisord.

Things to look at: max_requests in combo with max_requests_jitter. We should be able to do better with keep-alive requests, which needs changes in haproxy config.
Lots of corner cases have been fixed.

We should look at some changes to our stack, notably some simplifications and the use of uWSGI.


git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtPlatform/trunk@61028 3dfb626b-079f-4ab1-a9f6-4654e5377508