Only the BinaryUserRatings used for liking/favoriting ever returned the old rating. The generic UserRatings used for star ratings never has.
The analytics code was relying on that value being present and if it wasn't threw an exception:

  nti/analytics/resource_tags.py", line 112, in _note_rated
    timestamp = event.rating.timestamp
  AttributeError: 'NoneType' object has no attribute 'timestamp'"

This fixes test_note_rated in nti.appserver.tests.test_application_rating.TestApplicationRating.

I think these tests must not be running, probably due to the layer setup issues when the whole suite is run.



git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtPlatform/trunk@65268 3dfb626b-079f-4ab1-a9f6-4654e5377508