Add info to the SIGPROF handler and /stacktraces view to diagnose memory leaks.
This adds the `objgraph` dependency.

The /stacktraces view is only available in devmode sites. For all
others you need to use the SIGPROF handler.

Note that the growth stats from the first time this is done will not
be correct (they'll report growth since the process started).

This is a fairly expensive operation because it tracks through all the
objects in the heap (it already did that to report stacks, now it does
it more times), so don't do it often, only when manually diagnosing a
problem.



git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtPlatform/trunk@98189 3dfb626b-079f-4ab1-a9f6-4654e5377508